pyproject.toml
[project]
name = "new-bill-calculator"
version = "0.1.0"
description = ""
authors = [
    {name = "VinylStage",email = "mcpe9869@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (==2.3.3)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
